+++ piplicenses/tomli_bridge.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 # vim:fenc=utf-8 ff=unix ft=python ts=4 sw=4 sts=4 si et
 """
 pip-licenses.tomil_support
