+++ pyproject.toml
@@ -10,27 +10,27 @@ requires-python = ">=3.9"
 license = "MIT"
 license-files = ["LICENSE"]
 authors = [
-    {"name" = "raimon", "email" = "raimon49@hotmail.com"}, # 2018-2025 (retired)
-    {"name" = "Mr. Walls", "email" = "reactive-firewall@users.noreply.github.com"},
+	{"name" = "raimon", "email" = "raimon49@hotmail.com"}, # 2018-2025 (retired)
+	{"name" = "Mr. Walls", "email" = "reactive-firewall@users.noreply.github.com"},
 ]
 maintainers = [
-    {"name" = "Mr. Walls", "email" = "reactive-firewall@users.noreply.github.com"},
+	{"name" = "Mr. Walls", "email" = "reactive-firewall@users.noreply.github.com"},
 ]
 keywords = ["pip", "pypi", "package", "license", "check", "audit"]
 classifiers = [
-    "Development Status :: 5 - Production/Stable",
-    "Intended Audience :: Developers",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Programming Language :: Python :: 3.14",
-    "Topic :: System :: Systems Administration",
-    "Topic :: System :: System Shells",
-    "Typing :: Typed"
+	"Development Status :: 5 - Production/Stable",
+	"Intended Audience :: Developers",
+	"Programming Language :: Python :: 3",
+	"Programming Language :: Python :: 3 :: Only",
+	"Programming Language :: Python :: 3.9",
+	"Programming Language :: Python :: 3.10",
+	"Programming Language :: Python :: 3.11",
+	"Programming Language :: Python :: 3.12",
+	"Programming Language :: Python :: 3.13",
+	"Programming Language :: Python :: 3.14",
+	"Topic :: System :: Systems Administration",
+	"Topic :: System :: System Shells",
+	"Typing :: Typed"
 ]
 dependencies = [
     "prettytable >= 3.12.0",
@@ -40,16 +40,16 @@ scripts = { "pip-licenses" = "piplicenses.__main__:main" }
 
 [project.optional-dependencies]
 dev = [
-    "docutils>=0.19", # frontend.get_default_settings()
-    "wheel",
-    "mypy==1.19.1", # v1.19.1 is needed for python 3.9 support
-    "pip-tools",
-    "pypandoc",
-    "pytest-cov",
-    "pytest-runner",
-    "ruff",
-    "tomli-w",
-    "virtualenv>=21.1.0",
+	"docutils>=0.19", # frontend.get_default_settings()
+	"wheel",
+	"mypy==1.19.1", # v1.19.1 is needed for python 3.9 support
+	"pip-tools",
+	"pypandoc",
+	"pytest-cov",
+	"pytest-runner",
+	"ruff",
+	"tomli-w",
+	"virtualenv>=21.1.0",
 ]
 
 [tool.pip-tools.compile]
@@ -61,8 +61,9 @@ output-file = "dev-requirements.in"
 
 [project.urls]
 homepage = "https://github.com/raimon49/pip-licenses"
-releasenotes = "https://github.com/raimon49/pip-licenses/releases"
 issues = "https://github.com/raimon49/pip-licenses/issues"
+releasenotes = "https://github.com/raimon49/pip-licenses/releases"
+repository = "https://github.com/raimon49/pip-licenses.git"
 
 [tool.setuptools]
 py-modules = ["pip-licenses"]
@@ -83,30 +84,34 @@ target-version = "py39"
 
 [tool.ruff.lint]
 extend-select = [
-    "ANN",    # flake8-annotations
-    "B",      # flake8-bugbear
-    "C4",     # flake8-comprehensions
-    "FA",     # flake8-future-annotations
-    "ISC",    # flake8-implicit-str-concat
-    "LOG",    # flake8-logging
-    "G",      # flake8-logging-format
-    "PIE",    # flake8-pie
-    "PYI",    # flake8-pyi
-    "SIM",    # flake8-simplify
-    "FLY",    # flynt
-    "I",      # isort
-    "PERF",   # Perflint
-    "PGH",    # pygrep-hooks
-    "UP",     # pyupgrade
-    "FURB",   # refurb
-    "RUF",    # Ruff-specific rules
+	"ANN",    # flake8-annotations
+	"B",      # flake8-bugbear
+	"C4",     # flake8-comprehensions
+	"FA",     # flake8-future-annotations
+	"ISC",    # flake8-implicit-str-concat
+	"LOG",    # flake8-logging
+	"G",      # flake8-logging-format
+	"PIE",    # flake8-pie
+	"PYI",    # flake8-pyi
+	"SIM",    # flake8-simplify
+	"FLY",    # flynt
+	"I",      # isort
+	"PERF",   # Perflint
+	"PGH",    # pygrep-hooks
+	"UP",     # pyupgrade
+	"FURB",   # refurb
+	"RUF",    # Ruff-specific rules
 ]
-ignore = [
-    "C401",   # Unnecessary generator (rewrite as a set comprehension)
-    "F841",   # Local variable is assigned to but never used
-    "SIM300", # Yoda condition detected
-    "RUF005", # Consider unpacking instead of concatenation
-    "RUF021", # Parenthesize `a and b` expressions when chaining `and` and `or` together, to make the precedence clear
+extend-ignore = [
+	"C401",   # Unnecessary generator (rewrite as a set comprehension)
+	"F841",   # Local variable is assigned to but never used
+	"SIM300", # Yoda condition detected
+	"RUF005", # Consider unpacking instead of concatenation
+	"RUF012", # Unnecessary noise about Python language (In Python defaults are always mutable)  # See https://github.com/astral-sh/ruff/issues/5243
+	"RUF021", # Parenthesize `a and b` expressions when chaining `and` and `or` together, to make the precedence clear
+	"RUF013", # Ignore PEP484 until GHI-316 is resolved
+	"FA100",  # DEPRECATED in v6.0+  # See new
+	"N999",   # Ignore until GHI-316 is resolved
 ]
 isort.known-first-party = ["piplicenses"]
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?)(_{2,2})?)$"  # See https://regex101.com/r/U0tpDT/2
